refactor(dashboard): delete constants.ts and inline BillingActions
constants.ts duplicated FEATURE_IDS from billing-plans.ts with a
comment saying 'to avoid importing from @epicenter/api'—but we
already depend on that package. Import directly instead.

BillingActions had one consumer and its openPortal() was duplicated
in +page.svelte after the trial CTA move. Inline the top-up mutation
and buttons into the page, eliminating the duplication and the
component file.

Net -28 lines, 2 files deleted.
